fixed the structure
