--- __init__.py
@@ -1,4 +1,0 @@
-import Advanced_board_elements
-import Simple_board_elements
-import Fundamentals
-import visual1_many
