+++ app_code/board_elements/Fundamentals.py
@@ -1,6 +1,11 @@
 from __future__ import annotations
+import os
+import sys
 
-from exceptions import *
+root_folder = os.path.abspath(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+sys.path.append(root_folder)
+
+from exceptions.exceptions import *
 
 
 class Pin:
