+++ app_code/junk/old code parts.py
