+++ app_code/tests/TestFund.py
@@ -1,8 +1,13 @@
 from unittest import TestCase, main
-from Fundamentals import Pin, InputPin, OutputPin
-from Fundamentals import Board
-from Fundamentals import BaseCircuitElement
-from exceptions import *
+import os
+import sys
+
+root_folder = os.path.abspath(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+sys.path.append(root_folder)
+from board_elements.Fundamentals import Pin, InputPin, OutputPin
+from board_elements.Fundamentals import Board
+from board_elements.Fundamentals import BaseCircuitElement
+from exceptions.exceptions import *
 class TestFund(TestCase):
     def test_BCE(self):
