+++ app_code/tests/Test_SBE.py
@@ -1,6 +1,12 @@
 from unittest import TestCase, main
-from  Fundamentals import Board, BaseCircuitElement
-from Simple_board_elements import *
+import os
+import sys
+
+root_folder = os.path.abspath(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+sys.path.append(root_folder)
+
+from  board_elements.Fundamentals import Board, BaseCircuitElement
+from board_elements.SimpleBoardElements import *
 
 class Test_SBE(TestCase):
     """
