+++ app_code/visuals/textures/and.png
