+++ app_code/visuals/textures/buffer.png
