+++ app_code/visuals/textures/light_bulb.png
