+++ app_code/visuals/textures/light_bulb_shine.png
