+++ app_code/visuals/textures/low_constant.png
