--- tests/__init__.py
@@ -1,2 +1,0 @@
-import sys
-sys.path.insert(0,'..')
